src/requirements.txt
alembic
SQLAlchemy
pyTelegramBotAPI
requests[socks]
emoji
pymysql
python-i18n[YAML]
aiohttp
